Automate The Date Value with the use of API Github
This function retrieves the SHA value of the latest commit in the "experimental_7.0" branch of the Proton repository from Github API. It then extracts the date of the commit and formats it as "7.0-YYYYMM". Finally, it replaces any hyphens in the date with dots and sets it as the value of the pkgver variable.