pacutils: Add PrettyProgressBar Config Option
The pacman PKGBUILD we provide has a patch that enables a unicode progress bar when built with the
necessary options. When used with upstream pacutils however, pacconf will print out warnings about
"PrettyProgressBar" config option not existing because of the way it parses configs manually.

For QoL reasons, I have made a patch to add that config option to pacconf directly so it doesn't complain
when used with PrettyProgressBar'ed pacman

Signed-off-by: Eric Naim <dnaim@cachyos.org>