pacman: Build with PrettyProgressBar (#377)
* pacutils: Add PrettyProgressBar Config Option

The pacman PKGBUILD we provide has a patch that enables a unicode progress bar when built with the
necessary options. When used with upstream pacutils however, pacconf will print out warnings about
"PrettyProgressBar" config option not existing because of the way it parses configs manually.

For QoL reasons, I have made a patch to add that config option to pacconf directly so it doesn't complain
when used with PrettyProgressBar'ed pacman

Signed-off-by: Eric Naim <dnaim@cachyos.org>

* pacman: Enable PrettyProgressBar

Builds pacman with PrettyProgressBar functionality. This option, when enabled, will render the progress bar using
specifing unicode characters so it looks "pretty". It is disabled by default for compatibility reasons (some terminals
cannot render unicode characters nicely, and not all fonts have those specific unicode characters). I've moved relevant
patch to the main source array to facilitate this change.

Signed-off-by: Eric Naim <dnaim@cachyos.org>

---------

Signed-off-by: Eric Naim <dnaim@cachyos.org>