checker: Pass 3rd argument for GET requests
We have nounset in the script to ensure that both GITHUB_TOKEN and
GITHUB_REPO is set. However, this bites us in the ass because it will
fail in github_api when we're doing GET requests because we don't pass a
third argument.

By passing an arbitrary argument, we can fix the script and make it work
gracefully. See [1] for a working example of the script.

[1] https://github.com/1Naim/CachyOS-PKGBUILDS/issues/1

Signed-off-by: Eric Naim <dnaim@cachyos.org>