zfs-utils: Import from AUR
Up until now, we have been relying on the AUR package for zfs-utils.
This keeps biting us in the arse however since it's not in-sync with
whatever patches we are using. It is not the first time that issues like
[1] show up in CachyOS due to this.

Maintain our own PKGBUILD of zfs-utils to make sure that the userspace
tools and kernel module are always the same version and no breakage from
version mismatches can occur.

Signed-off-by: Eric Naim <dnaim@cachyos.org>