bcachefs: use UUID for rootfs in kernel cmd line
As of September 2025, CachyOS default install with a single-drive
bcachefs filesystem for root works. However, with more than one disk in
the bcachefs filesystem, sdboot-manage uses the colon-delimited list of
devices for root= in the kernel command line, which fails. This patch
will find the bcachefs filesystem UUID, and use that for root= instead.