proton,wine: version 10.0-20251222
With the current setup, proton files are under `/usr` and the runtime doesn't quite like it because it can't mount the `/usr`
location in the container, so I moved them to `/opt`, and placed a description file in the old install location to point to
the new installation. Steam works just fine like this, it is part of their specification actually, but no other launcher
than Steam and Rare supports this to my knowledge.